test: exercise what makes getOpenTabCount correct, and say what it counts
The two tests stubbed dockLayout.find with a fake that discarded the callback's
return value and never looked at the filter argument, so neither property the
implementation depends on was checked. Both of the mutations that break it in
production passed green: returning true from the callback halts rc-dock's walk
at the top-level box it offers first (the #253 bug the isTab guard exists for),
and Filter.Tab in place of Filter.AnyTab enters no box at all. Either one makes
the count 0, which makes getOpenTabCountSync() <= 1 permanently true and drops
"Move tab to new window" from every tab whenever a second window is open.

The stub now walks the way Algorithm.find does — offering the top-level box
first, stopping on a truthy callback, and answering nothing unless asked with
Filter.AnyTab — and a third test captures the filter argument directly. Mutated
both ways against the new tests: `return true` reddens two, and Filter.Tab
reddens three including the captured-filter assertion (expected 1 to be 121).

The doc said the count covers tabs "docked or floated". Filter.AnyTab is
Tab|Docked|Floated|Windowed|Max, so it also walks the window and max boxes; the
same wording had been copied onto the public PapiDockLayout declaration, so both
are corrected.

Also moves the two context-menu localization helpers into their own module.
They were exported from the overlay component for the previous commit, and that
component's module is part of papi.d.ts, so they reach the generated types
either way — the dts bundler follows imports, and relocating does not change
that. They therefore carry @experimental, like every other new public item here.
