+++ src/renderer/components/docking/platform-dock-layout-storage.util.ts
@@ -629,7 +629,8 @@ export function getAllWebViewDefinitions(dockLayout: DockLayout): WebViewDefinit
   const webViewDefinitions: WebViewDefinition[] = [];
 
   // Always return false from the callback so rc-dock visits every tab instead of stopping at the
-  // first match. Filter.AnyTab traverses both docked and floated panels.
+  // first match. Filter.AnyTab is Tab|Docked|Floated|Windowed|Max, so the walk enters every box the
+  // layout has — not only the docked and floating ones.
   dockLayout.find((item) => {
     // Still have to check isTab because of a bug https://github.com/ticlo/rc-dock/pull/253
     if (!isTab(item)) return false;
@@ -652,7 +653,7 @@ export function getAllWebViewDefinitions(dockLayout: DockLayout): WebViewDefinit
  *
  * @param dockLayout The rc-dock dock layout React component ref. Used to perform operations on the
  *   layout
- * @returns The number of tabs open anywhere in the layout, docked or floated
+ * @returns The number of tabs open anywhere in the layout, whatever its docking state
  */
 export function getOpenTabCount(dockLayout: DockLayout): number {
   let tabCount = 0;
